moved process waiting to cleanup class, user log was not shown
