+++ src/lib/cleanup.js
@@ -10,6 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
+const execa = require('execa')
 const aioLogger = require('@adobe/aio-lib-core-logging')('@adobe/aio-cli-plugin-app:cleanup', { provider: 'debug' })
 
 /** @private */
@@ -22,6 +23,25 @@ class Cleanup {
     this.resources.push(async () => {
       aioLogger.debug(message)
       await func()
+    })
+  }
+
+  async wait () {
+    if (this.resources.length < 1) {
+      const dummyProc = execa('node')
+      this.add(async () => await dummyProc.kill(), 'stopping sigint waiter...')
+    }
+    // bind cleanup function
+    process.on('SIGINT', async () => {
+      try {
+        await this.run()
+        aioLogger.info('exiting!')
+        process.exit(0) // eslint-disable-line no-process-exit
+      } catch (e) {
+        aioLogger.error('unexpected error while cleaning up!')
+        aioLogger.error(e)
+        process.exit(1) // eslint-disable-line no-process-exit
+      }
     })
   }
