+++ src/lib/run-dev.js
@@ -26,7 +26,6 @@ const deployActions = require('./deploy-actions')
 const actionsWatcher = require('./actions-watcher')
 
 const utils = require('./app-helper')
-const execa = require('execa')
 const { run: logPoller } = require('./log-poller')
 
 /** @private */
@@ -57,24 +56,7 @@ async function runDev (args = [], config, options = {}, log = () => {}) {
   devConfig.envFile = '.env'
 
   const cleanup = new Cleanup()
-  let needsProcessWaiter = true
   let defaultBundler = null
-
-  // bind cleanup function
-  process.on('SIGINT', async () => {
-    // in case app-scripts are eventually turned into a lib:
-    // - don't exit the process, just make sure we get out of waiting
-    // - unregister sigint and return properly (e.g. not waiting on stdin.resume anymore)
-    try {
-      await cleanup.run()
-      log('exiting!')
-      process.exit(0) // eslint-disable-line no-process-exit
-    } catch (e) {
-      aioLogger.error('unexpected error while cleaning up!')
-      aioLogger.error(e)
-      process.exit(1) // eslint-disable-line no-process-exit
-    }
-  })
 
   try {
     // Build Phase - actions
@@ -83,7 +65,6 @@ async function runDev (args = [], config, options = {}, log = () => {}) {
         const { config: localConfig, cleanup: localCleanup } = await runLocalRuntime(config, log, options.verbose)
         devConfig = localConfig
         cleanup.add(() => localCleanup(), 'cleaning up runDevLocal')
-        needsProcessWaiter = false
       } else {
         // check credentials
         rtLibUtils.checkOpenWhiskCredentials(devConfig)
@@ -142,7 +123,6 @@ async function runDev (args = [], config, options = {}, log = () => {}) {
           const { url, cleanup: serverCleanup } = result
           frontEndUrl = url
           cleanup.add(() => serverCleanup(), 'cleaning up serve...')
-          needsProcessWaiter = false
         }
       }
     }
@@ -156,17 +136,8 @@ async function runDev (args = [], config, options = {}, log = () => {}) {
     if (config.app.hasBackend && fetchLogs) {
       const { cleanup: pollerCleanup } = await logPoller(devConfig)
       cleanup.add(() => pollerCleanup(), 'cleaning up log poller...')
-      needsProcessWaiter = false
     }
-
-    // if there is no process waiting (for example OpenWhisk, etc)
-    // we need to explicitly wait for CTRL-C with a dummy process
-    if (needsProcessWaiter) {
-      const dummyProc = execa('node')
-      cleanup.add(async () => await dummyProc.kill(), 'stopping sigint waiter...')
-    }
-
-    log('press CTRL+C to terminate dev environment')
+    cleanup.wait()
   } catch (e) {
     aioLogger.error('unexpected error, cleaning up...')
     await cleanup.run()
